requirements.txt
# Core Dependencies
requests>=2.28.0
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
pyyaml>=6.0.0
rich>=12.0.0
click>=8.1.0

# Security
cryptography>=38.0.0
pyOpenSSL>=22.0.0
certifi>=2022.12.7
jwt>=1.3.1

# Database & Caching
redis>=4.4.0
sqlalchemy>=1.4.0
alembic>=1.8.0

# Async & Concurrency
asyncio>=3.4.3
aioredis>=2.0.0
concurrent-log-handler>=0.9.20

# UI & Reporting
jinja2>=3.1.0
markdown>=3.4.0
pygments>=2.13.0
reportlab>=3.6.0

# Testing
pytest>=7.2.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Development
black>=22.12.0
mypy>=0.991
flake8>=6.0.0
isort>=5.11.0

# Monitoring & Logging
prometheus-client>=0.15.0
structlog>=22.3.0
python-json-logger>=2.0.4

# Web & API
fastapi>=0.88.0
uvicorn>=0.20.0
websockets>=10.4
httpx>=0.23.0

# Documentation
sphinx>=5.3.0
mkdocs>=1.4.0
mkdocs-material>=8.5.0

# Optional Features
pillow>=9.3.0
python-magic>=0.4.27
psutil>=5.9.0
